requirements.txt
opencv-python
rawpy
numpy
pillow
exifread
tqdm
pandas